% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{algo_leverage}
\alias{algo_leverage}
\title{Algorithm leveraging function}
\usage{
algo_leverage(y, X, size, draws, method = "uniform")
}
\arguments{
\item{y}{The response}

\item{X}{The covariate}

\item{size}{subsample size of leveraging}

\item{draws}{Number of draws/repetition of leveraging}

\item{method}{The method to implement, can be weighted leveraging and uniform
leveraging. Default is "uniform". Details of the difference of the}
}
\value{
the leveraged coefficient that fits the linear regression
}
\description{
A function that implements algorithmic leveraging for linear regression using uniform and
leverage score based subsampling of rows.
}
\author{
Yixiao Lin
}
