% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.R
\name{CD}
\alias{CD}
\title{coordinate descent elastic net}
\usage{
CD(X, Y, lambda, alpha, beta, n, p, maxit, tol)
}
\arguments{
\item{X}{The covariate}

\item{Y}{The response}

\item{lambda}{Regularization parameter}

\item{alpha}{Elastic net parameter, 0<=alpha<=1}

\item{beta}{The initialization of the coeffient.}

\item{n}{numbers of rows of X}

\item{p}{numbers of rows of Y}

\item{maxit}{Maximum Number of Iteration.}

\item{tol}{stop criterion. If the difference between the current result
and result from the last iteration is less than the Tolerance, then we say the algorithm converges}
}
\description{
coordinate descent elastic net
}
