% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.R
\name{elnet_coord}
\alias{elnet_coord}
\title{Elastic Net}
\usage{
elnet_coord(y, X, beta, alpha, lambda, maxit = 10000, tol = 1e-06)
}
\arguments{
\item{y}{The response}

\item{X}{The covariate}

\item{beta}{The initialization of the coeffient. The default value is 0.}

\item{alpha}{Elastic net parameter, 0<=alpha<=1}

\item{lambda}{Regularization parameter}

\item{maxit}{Maximum Number of Iteration. The default value is 10000.}

\item{tol}{stop criterion. If the difference between the current result
and result from the last iteration is less than the Tolerance, then we say the algorithm converges
The default value is 1e-6}
}
\value{
The solution that fits elastic net to the given data using
coordinate descent algorithm.
}
\description{
A function that fits elastic net to the given data using coordinate descent algorithm.
}
\author{
Yixiao Lin
}
