% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{f.iter}
\alias{f.iter}
\title{iterative method for GS, Jacobi and Parallel Jacobi}
\usage{
f.iter(A, b, x0, f, maxit)
}
\arguments{
\item{A}{Linear System of interest}

\item{b}{Target Vector}

\item{x0}{Initial guess of the solution}

\item{f}{function used for solving Ax=b. There are 3 options: "Gauss" "Jacobi"  and "Parallel Jacobi"}

\item{maxit}{Maximum Number of Iteration}
}
\description{
iterative method for GS, Jacobi and Parallel Jacobi
}
