% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{getL}
\alias{getL}
\title{obtain the lower triangular matrix of A}
\usage{
getL(A)
}
\arguments{
\item{A}{Linear System of interest}
}
\value{
the lower triangular matrix of A
}
\description{
obtain the lower triangular matrix of A
}
\author{
Yixiao Lin
}
