% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{getU}
\alias{getU}
\title{obtain the upper triangular matrix of A}
\usage{
getU(A)
}
\arguments{
\item{A}{Linear System of interest}
}
\value{
the  upper triangular matrix of A
}
\description{
obtain the upper triangular matrix of A
}
\author{
Yixiao Lin
}
