% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{Jac}
\alias{Jac}
\title{Jacobi Method}
\usage{
Jac(A, b, x)
}
\arguments{
\item{A}{Linear System of interest}

\item{b}{Target Vector}

\item{x}{Initial guess of the solution}
}
\description{
Jacobi Method
}
