% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{lev_subsample}
\alias{lev_subsample}
\title{subsample data with draws repetition, size r, probability Pi}
\usage{
lev_subsample(y, X, r, Pi, draws)
}
\arguments{
\item{y}{The response}

\item{X}{The covariate}

\item{r}{subsample size of leveraging}

\item{Pi}{the probability vector to select r sample out of n}

\item{draws}{Number of draws/repetition of leveraging}
}
\value{
the leveraged coefficient that fits the linear regression
}
\description{
subsample data with draws repetition, size r, probability Pi
}
\author{
Yixiao Lin
}
