% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{solve_ols}
\alias{solve_ols}
\title{Solve linear system}
\usage{
solve_ols(A, b, x0, Cores = 1, parallel = F, method = "Gauss", maxit = 10000)
}
\arguments{
\item{A}{Linear System of interest}

\item{b}{Target Vector}

\item{x0}{Initial guess of the solution}

\item{Cores}{number of cores for parallel Jacobi.}

\item{parallel}{TRUE if the method is parallel Jacobi. The default is FALSE.}

\item{method}{"Gauss" or "Jacobi". The default is "Gauss"}

\item{maxit}{Maximum Number of Iteration}
}
\value{
the solution to the linear system
}
\description{
A function that solves linear system using either Gauss-Seidel or Jacobi
}
\author{
Yixiao Lin
}
